requirements.txt
python-telegram-bot==20.8
python-dotenv
qrcode[pil]
